+++ tests/SingleDepthValidatorTest.php
@@ -2,7 +2,7 @@
 
 
 use Briedis\ApiBuilder\Exceptions\InvalidStructureException;
-use Briedis\ApiBuilder\Items\Float;
+use Briedis\ApiBuilder\Items\Decimal;
 use Briedis\ApiBuilder\Items\Integer;
 use Briedis\ApiBuilder\StructureBuilder;
 use Briedis\ApiBuilder\StructureValidator;
@@ -73,7 +73,7 @@ class SingleDepthValidatorTest extends PHPUnit_Framework_TestCase{
 			$fields = $e->getBadFields();
 
 			$this->assertInstanceOf(get_class(new Integer), $fields['id'], 'Correct field is invalid');
-			$this->assertInstanceOf(get_class(new Float), $fields['decimal'], 'Correct field is invalid');
+			$this->assertInstanceOf(get_class(new Decimal), $fields['decimal'], 'Correct field is invalid');
 			$this->assertEquals(2, count($fields), 'Bad parameter count matches');
 		}
